backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
email-validator>=2.0.0
openai>=1.12.0
httpx>=0.27.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
